backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pypdfium2==5.7.1
python-dotenv==1.0.1
python-multipart==0.0.27
sniffio==1.3.1
starlette==1.0.0
supabase
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
